Correct client id handling in Shim
The way we currently handle client ids in VarD is pretty broken, since
if a client connects to multiple Raft servers (say, in the case of a
failover) it will have different ids! This new shim fixes this, since
it expects clients to send their ids with every request and associates
sockets with ids on this basis. Clients are now responsible for
ensuring id uniqueness.